Finished implementing SSAO + backface SSAO (#2)
* Added base render stage for forward passes

* Changed VS and PS names for forward render stage

* Implemented backface rendering - but default materials break other rendering stages

* Optimized UBO setting (no update if data hasn't changed)

* Added more control over how to set default material in material store

* Implemented backface depth pass

* Fixed geometry for deferred spotlights and added option to disable diffuse texturing

* Finished SSAO implementation

+++ Wasabi/Materials/WMaterial.h
@@ -178,7 +178,7 @@ public:
 	 * @param  len     Length of data, in bytes
 	 * @return         Error code, see WError.h
 	 */
-	WError SetVariableData(const char* varName, void* data, int len);
+	WError SetVariableData(const char* varName, void* data, size_t len);
 
 	/**
 	 * Sets a texture in the bound effect.
